Change from time chart pivot hour test to time data processor test
