--- test/time_chart_test/time_chart_pivot_hour_test.dart
@@ -1,186 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:time_chart/src/chart.dart';
-import 'package:time_chart/time_chart.dart';
-
-import '../utils/chart_state_utils.dart';
-
-void main() {
-  group('Time chart pivot hours test', () {
-    testWidgets('merge if has overlapping time', (tester) async {
-      await tester.pumpWidget(MaterialApp(
-        home: TimeChart(
-          data: [
-            DateTimeRange(
-              start: DateTime(2021, 2, 24, 23, 15),
-              end: DateTime(2021, 2, 25, 7, 30),
-            ),
-            DateTimeRange(
-              start: DateTime(2021, 2, 22, 1, 55),
-              end: DateTime(2021, 2, 22, 9, 12),
-            ),
-          ],
-          viewMode: ViewMode.monthly,
-        ),
-      ));
-      final ChartState chartState = getChartState(tester);
-
-      expect(chartState.topHour, 22);
-      expect(chartState.bottomHour, 10);
-    });
-
-    testWidgets('compare empty space for setting pivot hours', (tester) async {
-      await tester.pumpWidget(MaterialApp(
-        home: TimeChart(
-          data: [
-            DateTimeRange(
-              start: DateTime(2021, 2, 2, 18, 42),
-              end: DateTime(2021, 2, 2, 21, 22),
-            ),
-            DateTimeRange(
-              start: DateTime(2021, 2, 2, 11, 39),
-              end: DateTime(2021, 2, 2, 18, 2),
-            ),
-            DateTimeRange(
-              start: DateTime(2021, 2, 2, 4, 52),
-              end: DateTime(2021, 2, 2, 9, 0),
-            ),
-            DateTimeRange(
-              start: DateTime(2021, 2, 1, 22, 12),
-              end: DateTime(2021, 2, 2, 3, 30),
-            ),
-          ],
-          viewMode: ViewMode.monthly,
-        ),
-      ));
-      final ChartState chartState = getChartState(tester);
-
-      expect(chartState.topHour, 11);
-      expect(chartState.bottomHour, 9);
-    });
-
-    testWidgets('compare and merge time for setting pivot hours',
-        (tester) async {
-      await tester.pumpWidget(MaterialApp(
-        home: TimeChart(
-          data: [
-            DateTimeRange(
-              start: DateTime(2021, 2, 2, 14, 42),
-              end: DateTime(2021, 2, 2, 21, 22),
-            ),
-            DateTimeRange(
-              start: DateTime(2021, 2, 2, 11, 39),
-              end: DateTime(2021, 2, 2, 16, 2),
-            ),
-            DateTimeRange(
-              start: DateTime(2021, 2, 2, 2, 52),
-              end: DateTime(2021, 2, 2, 9, 0),
-            ),
-            DateTimeRange(
-              start: DateTime(2021, 2, 1, 22, 12),
-              end: DateTime(2021, 2, 2, 3, 30),
-            ),
-          ],
-          viewMode: ViewMode.weekly,
-        ),
-      ));
-      final ChartState chartState = getChartState(tester);
-
-      expect(chartState.topHour, 11);
-      expect(chartState.bottomHour, 9);
-    });
-
-    testWidgets('default pivot hours is used if there are no space',
-        (tester) async {
-      await tester.pumpWidget(MaterialApp(
-        home: TimeChart(
-          data: [
-            DateTimeRange(
-              start: DateTime(2021, 2, 2, 17, 42),
-              end: DateTime(2021, 2, 2, 22, 22),
-            ),
-            DateTimeRange(
-              start: DateTime(2021, 2, 2, 8, 39),
-              end: DateTime(2021, 2, 2, 18, 2),
-            ),
-            DateTimeRange(
-              start: DateTime(2021, 2, 2, 2, 52),
-              end: DateTime(2021, 2, 2, 9, 0),
-            ),
-            DateTimeRange(
-              start: DateTime(2021, 2, 1, 22, 12),
-              end: DateTime(2021, 2, 2, 3, 30),
-            ),
-          ],
-          viewMode: ViewMode.monthly,
-        ),
-      ));
-      final ChartState chartState = getChartState(tester);
-
-      expect(chartState.topHour, chartState.widget.defaultPivotHour);
-      expect(chartState.bottomHour, chartState.widget.defaultPivotHour);
-    });
-
-    testWidgets(
-        'set both pivot hours to 12 AM if both pivot hours are the same',
-        (tester) async {
-      await tester.pumpWidget(MaterialApp(
-        home: TimeChart(
-          data: [
-            DateTimeRange(
-              start: DateTime(2021, 12, 17, 3, 12),
-              end: DateTime(2021, 12, 18, 2, 30),
-            ),
-          ],
-          viewMode: ViewMode.monthly,
-        ),
-      ));
-      final ChartState chartState = getChartState(tester);
-
-      expect(chartState.topHour, chartState.widget.defaultPivotHour);
-      expect(chartState.bottomHour, chartState.widget.defaultPivotHour);
-    });
-
-    testWidgets(
-        'custom defaultPivotHour parameter is used if time range is fully visible',
-        (tester) async {
-      const pivotHour = 2;
-      await tester.pumpWidget(MaterialApp(
-        home: TimeChart(
-          defaultPivotHour: pivotHour,
-          data: [
-            DateTimeRange(
-              start: DateTime(2021, 12, 17, 3, 12),
-              end: DateTime(2021, 12, 18, 2, 30),
-            ),
-          ],
-          viewMode: ViewMode.monthly,
-        ),
-      ));
-      final ChartState chartState = getChartState(tester);
-
-      expect(chartState.topHour, pivotHour);
-      expect(chartState.bottomHour, pivotHour);
-    });
-
-    testWidgets(
-        'all data is not changed to next day if the bottom pivot hour is 12 AM',
-        (tester) async {
-      final rowData = DateTimeRange(
-        start: DateTime(2021, 12, 18, 22, 12),
-        end: DateTime(2021, 12, 18, 23, 30),
-      );
-      await tester.pumpWidget(MaterialApp(
-        home: TimeChart(
-          data: [
-            rowData,
-          ],
-          viewMode: ViewMode.weekly,
-        ),
-      ));
-      final ChartState chartState = getChartState(tester);
-
-      expect(chartState.processedSleepData.first.end, rowData.end);
-    });
-  });
-}
